[Merge] Added validation for data requests.js

+++ rent-a-car/data/cars.js
@@ -5,6 +5,7 @@ let { ObjectId } = require('mongodb');
 module.exports = {
     async createCar(model, type, color, numberDoors, seatingCapacity, hourlyRate, availability, engineType){
       if (!model || !type || !color || !numberDoors || !seatingCapacity || !hourlyRate || !availability || !engineType) {
+        throw "All fields must be present";
         throw "None of the feilds should be empty.";
       }
       if(!hourlyRate.includes('$/hr')){
@@ -51,6 +52,9 @@ module.exports = {
   },
     
     async getCar(id){
+      if (!id) {
+        throw "Id field must be present";
+      }
       const carCollection = await cars()
       let parsedId = ObjectId(id);
       const car = await carCollection.findOne({_id: parsedId})
@@ -59,6 +63,9 @@ module.exports = {
     },
 
     async remove(id) {
+      if (!id) {
+        throw "Id field must be present";
+      }
       var id = ObjectId(id);
       const carCollection = await cars()
       const deletionInfo = await carCollection.deleteOne({ _id: id });
@@ -69,6 +76,8 @@ module.exports = {
     },
 
     async update(id,model, type, color, numberDoors, seatingCapacity, hourlyRate, availability, engineType){
+      if (!model || !type || !color || !numberDoors || !seatingCapacity || !hourlyRate || !availability || !engineType) {
+        throw "All fields must be present";
       if (!id || !model || !type || !color || !numberDoors || !seatingCapacity || !hourlyRate || !availability || !engineType) {
         throw "None of the feilds should be empty.";
       }
